Architecture and tokenizer files upload
A collection of files for the Neuromodulatory Control Networks architecture project. Included in this upload are the GPT2 tokenizer files utilized by the custom open source tokenizer, the architecture files for the NCN, a versatile training script that can also create a .log file which can be utilized by the plotppl.py script to plot perplexity drop over global steps.